Add support for exporting different rocket groups
To support groups the way rockets are stored was changed. All rocket
positions/directions are stored in one array (all positions are in one
array, all directions are in another). Each rocket is the index of the
position/direction vector in the corresponding array.

E.g:
 ROCKET_POSITIONS[SOME_ROCKET]; // This is the position of SOME_ROCKET
 ROCKET_DIRECTIONS[SOME_ROCKET]; // This is the direction of SOME_ROCKET

Groups are array's of rocket indices. E.g:
 ROCKET_POSITION[GROUP[0]]; // Position of first rocket in group
 ROCKET_DIRECTION[GROUP[0]]; // Direction of first rocket in group